2- Issue with blank selections in user maint: (#4407)
File changed: /template/buildScript/ToCompile/asiLogin.w
This issue was introduced with the latest version of asiLogin, when locating the advantzware.ini file.  The location routine uses the "current directory" (which when run is usually N:\admin\envadmin).  The locator attempts to find the advantzware.ini file, and is successful, recording that location as "..\advantzware.ini" - that is, one directory UP from where I am now.  Later, after the user has chosen his selected environment, the current directory is moved to the environment (for example, N:\Environments\Prod).  This makes the reference to "one up from where I am now" equivalent to N:\Environments.  Since the file is not in that location, the lookup in viewers/user.w fails, and the selection lists all display as blanks.
Specs:
The solution is to modify the asiLogin program to assign the cIniLoc variable to be the full pathname of the advantzware.ini file: N:\admin\advantzware.ini.  Testing on AWS and local env show that this resolves the problem.

Test Case:
Run advantzware, select any environment and database
Open N-U-3.
Open any user record.
Success if the 3 selection lists on the right side are populated.

Other changes in this file (related to DbLevelM variables) are for internal use, and support the limitation of database choices to match the environment chosen