68061-CEVersion=New-foil and window calc
Change to ConversionProcs.p to support handling of MSI UOM and (actually) likely support conversion with metric dimensions.  Also convert basis weight of override values to a weight per EA in order to replicate
Primary change in EstimateCalcProcs was to use the NEW conversion program since the legacy also does not support MSI UOM apparently.  Additionally, in Joe's report, Ink calculation in legacy is Net of Window.  NetWindowArea calculation was not taking place prior to ink consumption calculation.  That is changed as well.
Basis weight for Foil and Window pre-converted to Lbs/MSF to behave similarly to board.  Disregard the basis weight value and move to use the Sqin/Lb value instead.